+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   * Support setting and editing the `description` field on project CI
     variables.
   * Support sorting tags by version number.
+  * Support filtering commits by author.
 
 ## Changes